more fixes, make selfroles controled by config

+++ BugBot.py
@@ -99,8 +99,9 @@ async def on_command_error(ctx: commands.Context, error):
         # notify caller
         await ctx.send(":rotating_light: Something went wrong while executing that command :rotating_light:", delete_after=10)
 
-    await asyncio.sleep(10)
-    await ctx.message.delete()
+    if ctx.guild is not None:
+        await asyncio.sleep(10)
+        await ctx.message.delete()
 
 
 def extract_info(o):
